==> 06_ana_malaria_prevalence.Rmd <==
---
title: "Long-term trends in malaria prevalence"
author: "Linda Petutschnig"
date: "28 04 2023"
output: 
  html_document:
    theme: paper
    highlight: default
---

This script performs the following tasks:

- Pre-processes and manipulates the data to prepare it for analysis.
- Generates a graph illustrating the development of Pf-incidence from 2000 to 2020 for all locations.
- Identifies a shift in the general trend around 2013 and conducts an in-depth analysis of the change between 2013 and 2020.
- Integrates the Pf-incidence values with a hexagonal grid prepared in a separate script.
- Uses the percentage change in Pf-incidence between 2013 and 2020 as a risk indicator and visualizes it on a map.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load libraries

```{r malariaAtlas, message = FALSE, warning = FALSE}

library(sf) # working with spatial vector data (simple features)
library(raster) # working with raster data
library(dplyr) # tidy data manipulation
library(here) # getting rid of relative paths
library(tidyr) # tidying data
library(lubridate) # to work with dates
library(stringr) # to manipulate sting variables
library(ggplot2) # plotting
library(tsibble) # time series analysis
library(fable) # time series analysis
library(MetBrewer) # color palettes for plots
library(tmap) # # for map making

source(here("00_hexagonify.R")) # a self-written function to summarize raster cells or points into hexagonal polygons

```

## Prepare and inspect data

```{r data analysis with dplyr, message = FALSE, error = FALSE}

# Loads the hexagons
aoi_shape <- st_read(here("data/geopackages","AOI_hex5.gpkg"))

# Loads the points containing the malaria data
pf_00_20_points <- st_read(here("data/geopackages","pf_00_20_points.gpkg"))

# Adds an ID column to the points
pf_00_20_points <- pf_00_20_points %>% mutate(id = row_number())

# A Year column and a value column are needed for the analysis.
# To this end, the gather function from the dplyr package is used to transform the data from wide to long format.
# (Note: consider using pivot_longer instead of gather, as it is the new function with the same functionality.)
pf_long <- pf_00_20_points %>% 
  gather(Year, MalariaIncidence, Plasmodium..falciparum..Incidence..2000:X202206_Global_Pf_Incidence_Rate_2020) 

# The Year column currently has long strings like "Plasmodium..falciparum..Incidence..2000".
# Only the last four characters are needed, which represent the Year information. 
pf_long$Year <- str_sub(pf_long$Year,-4,-1)      

# Converts the Year column to date format
# Adds Jan 1st to each row to create a proper date.
pf_long$Year <- lubridate::as_date(paste0(pf_long$Year, "-01-01"))

# Converts the long format data table to a time series tibble using its "id" information added previosuly
# Every geographical point is now represented by 21 rows holding data for the years 2000 to 2020, sharing the same id.
pf_tsbl <- as_tsibble(pf_long, key = id, index = Year)

# Calculates the average malaria incidence per year for every point individually
avg <- pf_long %>%
  group_by(Year) %>% 
  summarise(avg_value = mean(MalariaIncidence, na.rm = TRUE))

# creating a time series plot per location as a line
ggplot() + 
  geom_line(data =  pf_long, aes(x = Year, y = MalariaIncidence, group =id), color = "darkseagreen", alpha = 0.1) +
  geom_line(data = avg, aes(x = Year, y = avg_value), color = "salmon", linewidth = 2) +
  labs(x = "Year", y = "Malaria Incidence", title = "Malaria Incidence Over Time") +
  theme_minimal(base_size = 14) 

ggplot(data = pf_long, aes(x = MalariaIncidence, fill = "Malaria Incidence")) +
  geom_density(alpha = 0.5) +
  labs(x = "Malaria Incidence", y = "Density", title = "Density Plot of Malaria Incidence") +
  theme_minimal(base_size = 14)

```

## Filter relevant timespan

```{r since 2013, message = FALSE, error = FALSE, eval= FALSE}

# As can be seen from the figure above, the overall malaria incidence has decreased drastically between 2000 and 2020. 
# However,it also reveals that there has been a shift in the trend: Since around 2013, numbers are not decreasing anymore. 
# To better reflect this trend, the years 2000 to 2012 are removed in the next step. 
pf_tsbl_recent <- pf_tsbl %>% 
  dplyr::filter(Year > "2012-01-01")

avg_recent <- avg %>% 
  dplyr::filter(Year > "2012-01-01") 

# This function calculates the percantage by which the malaria incidence value has changed from 2013 to 2020.
perc_avg_13 <- ((avg_recent$avg_value[8] - avg_recent$avg_value[1]) / avg_recent$avg_value[1]) * 100

# Plots the years 2013 and 2020
# ggplot()+ 
#   geom_line(data =  pf_tsbl_recent, aes(x = Year, y = MalariaIncidence, group =id), color = "darkseagreen", alpha = 0.1) +
#   geom_line(data = avg_recent, aes(x= Year, y = avg_value), color = "salmon", linewidth = 2) +
#   labs(x = "Year", y = "Malaria Incidence", title = "Malaria Incidence change since 2013") +
#   theme_minimal(base_size = 14) 

# Calculates the change rate in malaria incidence between 2013 and 2020 in every individual location.
# Filters the values for 2013 and 2020.
# The same is done for a comparison between 2000 and 2020. 
value_2000 <- pf_tsbl %>% 
  dplyr::filter(Year == "2000-01-01") %>% 
  dplyr::pull(MalariaIncidence)

value_2013 <- pf_tsbl %>% 
  dplyr::filter(Year == "2013-01-01") %>% 
  dplyr::pull(MalariaIncidence)

value_2020 <- pf_tsbl %>% 
  dplyr::filter(Year == "2020-01-01") %>% 
  dplyr::pull(MalariaIncidence)

# Calculates the percentage.
perc_change_13 <- ((value_2020 - value_2013) / value_2013) * 100 # Change between 2013 and 2020
perc_change_00 <- ((value_2020 - value_2000) / value_2000) * 100 # Change between 2000 and 2020

# Writes the change rate in malaria incidence into a new column that is added to pf_00_20_points
pf_00_20_points$perc_change_13 <- perc_change_13
pf_00_20_points$perc_change_00 <- perc_change_00

```

## Hexagonify

```{r hexa, message = FALSE, error = FALSE, eval= FALSE}

# The data that goes into the hexagonify_points function is only allowed to have one column with values (in addition to the geometry column).
# Thus, the column containing the malaria incidence values must be filtered first.
pf_recent_points <- pf_00_20_points[24]

# The change rate in malaria incidence is read into the hexagons
hexa_points <- function(points, polygons, fun){
  
  # Checks which points intersect with which hexagon (polygons)
  poly_point_intersect <- st_intersects(polygons, points)
  
  # A function that is applied to every point per polygon, it returns the mean of all point values in the polygons 
  hexagon_value <- sapply(poly_point_intersect, function(ids) {
    pts <- points [ids, ]
    return(fun(pts[[1]])) # Accesses the value in the first column
  })
  
  # Fills in the average per hexagon cell 
  return(hexagon_value)
}

# As of 11.07.23 the hotspots_classify function gave an error, because in the just loaded dataset, some edges have duplicate
# vertexes with other edges. This has not been a problem before. 
# Adding the bit below solves the issue. 
sf_use_s2(FALSE)

aoi_shape$pf_hexa = hexa_points(pf_recent_points, aoi_shape, mean)
aoi_shape$pf_hexa[is.infinite(aoi_shape$pf_hexa)] <- NA     
#aoi_shape$pf_hexa[aoi_shape$pf_hexa == 0] <- NA

# Writes the results to hard drive.
st_write(aoi_shape, here("data/geopackages","AOI_hex5_ind2.gpkg"), append = FALSE)

```

## Visualize the results in a map
```{r where, message = FALSE, error = FALSE, eval= FALSE}

# Sets the tmap mode to an interactive map
tmap_mode("view")

# Defines the color palette
pal = met.brewer("Gauguin", type = "continuous", direction = -1)

# Creates the map 
tm_shape(aoi_shape) +
  tm_polygons(col = "pf_hexa", palette = pal, popup.vars = c("Change 13-20: "="pf_hexa", id = "h3_index")) +
  tm_layout(title = "Change in malaria incidence between 2013 and 2020") + 
  tm_fill(col = '#BFE1F4') + 
  tm_borders(lwd = 1, col = "black")  

```
