==> 08_ana_access_to_healthcare.Rmd <==
---
title: "Data analysis: Access to healthcare"
author: "Linda Petutschnig"
date: "2023-06-14"
output: html_document
---

Two indicators are calculated that are both related to healthcare accessibility:  
  1: Average walking time to nearest healthcare facility  
  2: Number of people served by the same healthcare facility  
  
This script performs the following tasks:  

- Loads a pre-prepared hexagonal DGGS grid  

 
- For indicator 1, the hexagonify function is used to calculate the average walking time in minutes per hexagon.
- For indicator 2, the population per service area is calculated and mapped. To this end, the "worldpop_2020_constrained_100m" data is loaded from hard drive which has been prepared in another script. The service areas were calculated using the SAGA GIS algorithm "Accumulate costs" with the "Walking-only travel time to healthcare map without access to motorized transport" data as input. The service areas are now overlaid with the population data. In a last step the population per service area is disaggregated into the hexagon layer.

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(ft.shadow = FALSE)

```

## Load libraries
```{r prepare, message = FALSE, warning = FALSE}

 # Loads necessary packages
library(here) # to access data via relative paths
library(malariaAtlas) # to access data provided by the Malaria Atlas Project
library(afrihealthsites) # in use?
library(sf) # to manipulate vector features
library(terra) # to work with raster files
library(raster) # to work with raster files
library(tmap) # for map making
library(Rsagacmd)

source(here("00_hexagonify.R")) # a self-written function to summarize raster cells or points into hexagonal polygons

```

## Define extent and download data   
```{r accessibility get data, message = FALSE, warning = FALSE}

# Loads hexagon layer from database
aoi_shape <- st_read(here("data/geopackages","AOI_hex5.gpkg"))

catalog <- listData(datatype = "raster")
aoi_spatial <- as(aoi_shape, "Spatial")

# Loads the accessibility layer from 
accessibility_raster <- getRaster(surface = "Walking-only travel time to healthcare map without access to motorized transport", shp = aoi_spatial)

writeRaster(accessibility_raster, here("data/rasters","AOI_traveltime_foot_hf.tif"), overwrite = TRUE)

```

# Indicator 1: Walking time to nearest healthcare facility  
```{r mean walking time per hexagon, message = FALSE, warning = FALSE}

sf_use_s2(FALSE)
# Call the hexagonify function to aggregate the accessibility values into a hexagonal grid.
aoi_shape$access <- hexagonify(accessibility_raster, aoi_shape, mean)

# Map
# Defines the color palette for the map
pal <- c('#60b564','#fed400')

# Sets the map to an interactive mode
tmap_mode("view")

# Creates the map object
tm_shape(aoi_shape) +
  tm_polygons(col = "access", palette = pal, popup.vars = c("Minutes"="access")) +
  tm_borders(lwd = 0.2) + 
  tm_layout(title = "Mean access time") + 
  tm_fill(col = '#BFE1F4') + 
  tm_borders(lwd = 1, col = "black") 

#Write the hexagons file into the db
st_write(aoi_shape, here("data/geopackages","AOI_hex5_ind3.gpkg"), append = FALSE)

```

## Indicator 2: Number of people served by the same healthcare facility  
```{r crop pop, message = FALSE, warning = FALSE}

# Load population statistics (raster)
pop <- rast(here("data/rasters", "AOI_ppp_2020_unconst_UNadj.tif"))

# Crop population dataset. 
pop_aoi <- terra::crop(pop, vect(aoi_shape), mask = TRUE)

# Plot the cropped population statistics
plot(pop_aoi)

```
## Calculate service areas (cost allocation)
```{r automate accumulated costs with SAGA, message = FALSE, warning = FALSE, echo = FALSE}

# Read AOI from geopackage
aoi <- st_read(here("data/geopackages","AOI.gpkg"))

# Dissolve admin boundaries in AOI data into a single polygon
aoi_dissolve <- st_union(aoi)

# Write the dissolved AOI into a new geopackage file
st_write(aoi_dissolve, here("data/geopackages", "AOI_dissolve.gpkg"), append = FALSE)

# Fetch health facilities data using the afrihealthsites package
hf_drc <- afrihealthsites("democratic republic of congo", datasource='who', plot='sf')
hf_bdi <- afrihealthsites("burundi", datasource='who', plot='sf')
hf_rwa <- afrihealthsites("rwanda", datasource='who', plot='sf')
hf_uga <- afrihealthsites("uganda", datasource='who', plot='sf')

# Combine health facilities data for all countries
hf_all <- rbind(hf_drc, hf_bdi, hf_rwa, hf_uga) 

# Check which health facilities intersect with the dissolved AOI
hf_aoi_logical <- st_intersects(hf_all, st_make_valid(aoi_dissolve))
intersection_indices <- sapply(hf_aoi_logical, sum) > 0

# Filter health facilities that intersect with the AOI
hf_aoi <- hf_all[intersection_indices, ]

# Write the filtered health facilities to new geopackage file
st_write(hf_aoi, here("data/geopackages","hf_aoi.gpkg"), append = FALSE)

# Convert accessibility_raster to terra raster object to be used with SAGA
accessibility_raster <- rast(accessibility_raster)

# Set up the SAGA GIS integration
saga <- saga_gis(
  saga_bin = "C:\\saga-9.1.1_x64\\saga_cmd.exe",
  raster_backend = "terra",
  vector_backend = "sf")

# Perform the cost allocation algorithm using SAGA GIS. Calculate accumulated cost for each health facility location
hf_allocated_costs <- saga$grid_analysis$accumulated_cost(dest_points = hf_aoi, cost = accessibility_raster)

```

## Count population per service area and map
```{r total pop served by one healthcare facility get friction surface, message = FALSE, warning = FALSE}

# Transform the allocazed cost raster into a vector object
service_areas <- as.polygons(hf_allocated_costs$allocation) %>% 
  st_as_sf()

# Transform into raster data type so that it can be utilized by the hexagonify function
pop_aoi <- raster(pop_aoi)

# Perfom hexagonify function to calculate the population served by each healthcare facility. The function is performed not with hexagones this time, but with the service areas.
service_areas$pop_served <- hexagonify(pop_aoi, service_areas, sum)

# Map the population served per healthcare facility using tmap
pal <- c('#60b564','#fed400')
tm_shape(service_areas) +
  tm_polygons(col = "pop_served", palette = pal, popup.vars = c("Population served"="pop_served")) +
  tm_borders(lwd = 0.2) + 
  tm_layout(title = "Population per healthcare facility") + 
  tm_fill(col = '#BFE1F4') + 
  tm_borders(lwd = 1, col = "black") 

```

## Calculate population per healthcare facility per hexagon and map
```{r integrate with DGGS, message = FALSE, warning = FALSE}

# Converts the 'service_areas' vector object into a raster based on the 'pop_served' field,
# using the 'accessibility_raster' as a template
service_areas_rast = rasterize(vect(service_areas), rast(accessibility_raster), field = "pop_served") %>% 
  raster()

# Calculates the mean population served within each hexagon in 'aoi_shape', 
# based on the 'service_areas_rast' raster.
aoi_shape$pop_served_hex <- hexagonify(service_areas_rast, aoi_shape, mean)

# Creates a map of the 'aoi_shape' polygons, using the 'pop_served_hex' field for coloring
pal <- c('#60b564','#fed400')
tm_shape(aoi_shape) +
  tm_polygons(col = "pop_served_hex", palette = pal, popup.vars = c("Pop served"="pop_served_hex")) +
  tm_borders(lwd = 0.2) + 
  tm_layout(title = "Population per healthcare facility per hexagon") + 
  tm_fill(col = '#BFE1F4') + 
  tm_borders(lwd = 1, col = "black") 

# Writes the 'aoi_shape' object (including the 'pop_served_hex' field) into a new geopackage file
st_write(aoi_shape, here("data/geopackages", "AOI_hex5_ind4.gpkg"))

```
